+++ CHANGELOG.md
@@ -7,6 +7,11 @@ The format of this change log follows the advice given at [Keep a CHANGELOG](htt
 
 ## [Unreleased]
 
+###  Added
+
+- Add support for Rule in the following rules:
+  - keywords-in-logical-order
+
 ###  Changed
 
 - The source repository has been moved from vsiakka/gherkin-lint to